User BMI & some fixes

+++ database/migrations/2020_07_15_173020_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
                         $table->string('password');
                         $table->string('phone',20)->nullable();
                         $table->integer('age')->nullable();
-                        $table->integer('height')->nullable();
+                        $table->float('height',3,2)->nullable();
                         $table->string('sex');
                         $table->boolean('showUserName')->default(true);
                         $table->string('role')->default('User');
